Removed old Overlay
Removed old Overlay
Added new Overlay and open button
TODO add hotkey for show / hide
TODO add ability to focus window or focus DCS?
