+++ DCS-SR-Client/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:DCS_SR_Client"
-             StartupUri="MainWindow.xaml">
+             StartupUri="ClientWindow/MainWindow.xaml">
     <Application.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
@@ -24,6 +24,8 @@
                 <!-- change "BaseLight" to the theme you want -->
                 <ResourceDictionary
                     Source="pack://application:,,,/MahApps.Metro;component/Styles/Accents/BaseLight.xaml" />
+
+                <ResourceDictionary Source="Themes\Styles.xaml" />
 
             </ResourceDictionary.MergedDictionaries>
         </ResourceDictionary>
