+++ DCS-SR-Client/ClientWindow/InputBindingControl.xaml
@@ -1,0 +1,52 @@
+﻿<UserControl x:Class="Ciribob.DCS.SimpleRadio.Standalone.Client.UI.InputBindingControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:Ciribob.DCS.SimpleRadio.Standalone.Client.UI"
+             mc:Ignorable="d" 
+             d:DesignHeight="100" d:DesignWidth="300">
+
+    <Grid>
+
+
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="2*" />
+            <ColumnDefinition Width="2*" />
+            <ColumnDefinition Width="1*" />
+            <ColumnDefinition Width="1*" />
+            <ColumnDefinition Width="1*" />
+        </Grid.ColumnDefinitions>
+        <Grid.RowDefinitions>
+            <RowDefinition />
+            <RowDefinition />
+      
+        </Grid.RowDefinitions>
+
+
+        <Label Grid.Column="0" Grid.Row="0" x:Name="DeviceLabel" Content="Control"
+                           VerticalContentAlignment="Center" />
+        <TextBox Grid.Column="1" Grid.Row="0" x:Name="Device" TextAlignment="Center"
+                             VerticalContentAlignment="Center" IsEnabled="False" Text="None" />
+        <TextBox Grid.Column="2" Grid.Row="0" x:Name="DeviceText" TextAlignment="Center"
+                             VerticalContentAlignment="Center" IsEnabled="False" Text="None" />
+        <Button Grid.Column="3" Grid.Row="0" x:Name="DeviceButton" Content="Set " Click="Device_Click"
+                            Style="{DynamicResource SquareButtonStyle}" />
+        <Button Grid.Column="4" Grid.Row="0" x:Name="DeviceClear" Content="Clear"
+                            Click="DeviceClear_Click" Style="{DynamicResource SquareButtonStyle}" />
+
+        <Label Grid.Column="0" Grid.Row="1" x:Name="ModifierLabel" Content="Control"
+                           VerticalContentAlignment="Center" />
+        <TextBox Grid.Column="1" Grid.Row="1" x:Name="ModifierDevice" TextAlignment="Center"
+                             VerticalContentAlignment="Center" IsEnabled="False" Text="None" />
+        <TextBox Grid.Column="2" Grid.Row="1" x:Name="ModifierText" TextAlignment="Center"
+                             VerticalContentAlignment="Center" IsEnabled="False" Text="None" />
+        <Button Grid.Column="3" Grid.Row="1" x:Name="ModifierButton" Content="Set " Click="Modifier_Click"
+                            Style="{DynamicResource SquareButtonStyle}" />
+        <Button Grid.Column="4" Grid.Row="1" x:Name="ModifierButtonClear" Content="Clear"
+                            Click="ModifierClear_Click" Style="{DynamicResource SquareButtonStyle}" />
+
+    </Grid>
+
+
+</UserControl>
