+++ DCS-SR-Client/ClientWindow/RadioChannelConfigUI.xaml
@@ -1,0 +1,20 @@
+﻿<UserControl x:Class="Ciribob.DCS.SimpleRadio.Standalone.Client.UI.RadioChannelConfigUi"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:local="clr-namespace:Ciribob.DCS.SimpleRadio.Standalone.Client.UI"
+             mc:Ignorable="d"
+             Name="RadioChannelConfig"
+             d:DesignHeight="50" d:DesignWidth="200">
+    <Grid>
+        <ComboBox Grid.Column="1" Grid.Row="1" Name="ChannelSelector" SelectedValuePath="Content"
+                  SelectionChanged="ChannelSelector_SelectionChanged">
+            <ComboBoxItem>Left</ComboBoxItem>
+            <ComboBoxItem>Right</ComboBoxItem>
+            <ComboBoxItem>Both</ComboBoxItem>
+        </ComboBox>
+
+
+    </Grid>
+</UserControl>
