+++ DCS-SR-Client/DCS-SR-Client.csproj
@@ -136,24 +136,35 @@
     <Compile Include="ClientAudioProvider.cs" />
     <Compile Include="DejitterBuffer.cs" />
     <Compile Include="AppConfiguration.cs" />
-    <Compile Include="InputBindingControl.xaml.cs">
+    <Compile Include="ClientWindow\InputBindingControl.xaml.cs">
       <DependentUpon>InputBindingControl.xaml</DependentUpon>
     </Compile>
     <Compile Include="InputConfiguration.cs" />
     <Compile Include="InputDevice.cs" />
     <Compile Include="InputDeviceManager.cs" />
-    <Compile Include="RadioChannelConfigUI.xaml.cs">
+    <Compile Include="ClientWindow\RadioChannelConfigUI.xaml.cs">
       <DependentUpon>RadioChannelConfigUI.xaml</DependentUpon>
     </Compile>
+    <Compile Include="RadioOverlayWindow\RadioControlGroup.xaml.cs">
+      <DependentUpon>RadioControlGroup.xaml</DependentUpon>
+    </Compile>
     <Compile Include="RadioFilter.cs" />
+    <Compile Include="RadioOverlayWindow\RadioOverlay.xaml.cs">
+      <DependentUpon>RadioOverlay.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="RadioReceivingState.cs" />
+    <Compile Include="RadioSendingState.cs" />
     <Compile Include="RadioSyncServer.cs" />
     <Compile Include="Settings.cs" />
+    <Compile Include="Themes\LeftMarginMultiplierConverter.cs" />
+    <Compile Include="Themes\TreeViewItemExtensions.cs" />
     <Compile Include="UDPVoiceHandler.cs" />
-    <Page Include="InputBindingControl.xaml">
+    <Compile Include="RadioOverlayWindow\WindowHelper.cs" />
+    <Page Include="ClientWindow\InputBindingControl.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="MainWindow.xaml">
+    <Page Include="ClientWindow\MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
@@ -162,13 +173,25 @@
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="ClientSync.cs" />
-    <Compile Include="MainWindow.xaml.cs">
+    <Compile Include="ClientWindow\MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Page Include="RadioChannelConfigUI.xaml">
+    <Page Include="ClientWindow\RadioChannelConfigUI.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="RadioOverlayWindow\RadioControlGroup.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Include="RadioOverlayWindow\RadioOverlay.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Include="Themes\Styles.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
     </Page>
   </ItemGroup>
   <ItemGroup>
