+++ DCS-SR-Client/Properties/AssemblyInfo.cs
@@ -25,7 +25,7 @@ using System.Windows;
 //<UICulture>CultureYouAreCodingWith</UICulture> in your .csproj file
 //inside a <PropertyGroup>.  For example, if you are using US english
 //in your source files, set the <UICulture> to en-US.  Then uncomment
-//the NeutralResourceLanguage attribute below.  Update the "en-US" in
+//the NeutralResourceLanguage attribute below.  RepaintRadioStatus the "en-US" in
 //the line below to match the UICulture setting in the project file.
 
 //[assembly: NeutralResourcesLanguage("en-US", UltimateResourceFallbackLocation.Satellite)]
