+++ DCS-SR-Client/RadioOverlayWindow/RadioOverlay.xaml
@@ -1,0 +1,61 @@
+﻿<Window x:Class="Ciribob.DCS.SimpleRadio.Standalone.Overlay.RadioOverlayWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:Ciribob.DCS.SimpleRadio.Standalone.Overlay"
+   
+        Style="{x:Null}"
+        Topmost="True"
+        Opacity="1.0"
+
+        ResizeMode="CanResizeWithGrip"
+
+        WindowStyle="None"
+        AllowsTransparency="True"
+        Background="#444"
+
+        Name="myMainWindow"
+        Title="DCS-SimpleRadio" MinHeight="260" MinWidth="122" Height="270" Width="122">
+
+    <WrapPanel Style="{x:Null}"  Name="containerPanel" SizeChanged="containerPanel_SizeChanged" Loaded="Window_Loaded">
+
+        <WrapPanel.LayoutTransform>
+            <ScaleTransform x:Name="ApplicationScaleTransform"
+                            CenterX="0"
+                            CenterY="0"
+                            ScaleX="{Binding ElementName=myMainWindow, Path=ScaleValue}"
+                            ScaleY="{Binding ElementName=myMainWindow, Path=ScaleValue}" />
+        </WrapPanel.LayoutTransform>
+
+        <Button Margin="90,0,0,0" Click="Button_Minimise" Style="{StaticResource DarkStyle-Button}">
+            <Button.Content>
+                <Grid Width="10" Height="10">
+                    <TextBlock Text="0" FontFamily="Marlett" FontSize="14" VerticalAlignment="Center"
+                               HorizontalAlignment="Center" Padding="3.5,0,0,3" />
+                </Grid>
+            </Button.Content>
+        </Button>
+
+        <Button Margin="2,0,0,0" Click="Button_Close" Style="{StaticResource DarkStyle-Button}">
+            <Button.Content>
+                <Grid Width="10" Height="10">
+                    <TextBlock Text="r" FontFamily="Marlett" FontSize="14" VerticalAlignment="Center"
+                               HorizontalAlignment="Center" Padding="0,0,0,1" />
+                </Grid>
+            </Button.Content>
+        </Button>
+
+        <local:RadioControlGroup x:Name="radio1" />
+        <Separator Margin="0" Width="120" />
+        <local:RadioControlGroup x:Name="radio2" />
+        <Separator Margin="0" Width="120" />
+        <local:RadioControlGroup x:Name="radio3" />
+        <Separator Margin="0 2" Width="120" />
+
+
+        <Slider Style="{x:Null}"  x:Name="windowOpacitySlider" VerticalAlignment="Center" Margin="10,2,0,0" Width="90" Minimum="0.05"
+                Maximum="1.0" ValueChanged="windowOpacitySlider_ValueChanged" Value="1.0" ToolTip="Window Opacity" />
+
+    </WrapPanel>
+</Window>
