+++ DCS-SR-Client/packages.config
@@ -1,5 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-
 <packages>
   <package id="Costura.Fody" version="1.3.3.0" targetFramework="net452" developmentDependency="true" />
   <package id="Fody" version="1.29.4" targetFramework="net452" developmentDependency="true" />
@@ -12,4 +11,5 @@
   <package id="semver" version="1.1.2" targetFramework="net452" />
   <package id="SharpDX" version="3.0.2" targetFramework="net452" />
   <package id="SharpDX.DirectInput" version="3.0.2" targetFramework="net452" />
+  <package id="WPFThemes.DarkBlend" version="1.0.8" targetFramework="net452" />
 </packages>
