+++ DCS-SimpleRadioStandalone.sln
@@ -11,13 +11,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DCS-SimpleRadio Server", "D
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DCS-SR-Common", "DCS-SR-Common\DCS-SR-Common.csproj", "{CF8DA5C1-0105-4486-9E28-CB2321D8F122}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DCS-SR-Radio-Interface", "RadioOverlay\DCS-SR-Radio-Interface.csproj", "{3FA37CBC-5BE2-4945-802D-9CDD464AC476}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Installer", "Installer\Installer.csproj", "{57B6F9D9-C410-4EB7-8003-60F894F358A4}"
 	ProjectSection(ProjectDependencies) = postProject
 		{CFF33E05-1D37-4807-B1F7-5A1488D8EC4F} = {CFF33E05-1D37-4807-B1F7-5A1488D8EC4F}
 		{838BDB0B-5EB1-4C1E-9026-0A8842AC00C6} = {838BDB0B-5EB1-4C1E-9026-0A8842AC00C6}
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476} = {3FA37CBC-5BE2-4945-802D-9CDD464AC476}
 		{CF8DA5C1-0105-4486-9E28-CB2321D8F122} = {CF8DA5C1-0105-4486-9E28-CB2321D8F122}
 		{8579C7DA-206F-4705-B073-D3CC946DC8EC} = {8579C7DA-206F-4705-B073-D3CC946DC8EC}
 	EndProjectSection
@@ -141,30 +138,6 @@ Global
 		{CF8DA5C1-0105-4486-9E28-CB2321D8F122}.Release|x64.Build.0 = Release|Any CPU
 		{CF8DA5C1-0105-4486-9E28-CB2321D8F122}.Release|x86.ActiveCfg = Release|Any CPU
 		{CF8DA5C1-0105-4486-9E28-CB2321D8F122}.Release|x86.Build.0 = Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Debug|x64.Build.0 = Debug|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Debug|x86.Build.0 = Debug|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Full Release|Any CPU.ActiveCfg = Full Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Full Release|Any CPU.Build.0 = Full Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Full Release|Mixed Platforms.ActiveCfg = Full Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Full Release|Mixed Platforms.Build.0 = Full Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Full Release|x64.ActiveCfg = Full Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Full Release|x64.Build.0 = Full Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Full Release|x86.ActiveCfg = Full Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Full Release|x86.Build.0 = Full Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Release|x64.ActiveCfg = Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Release|x64.Build.0 = Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Release|x86.ActiveCfg = Release|Any CPU
-		{3FA37CBC-5BE2-4945-802D-9CDD464AC476}.Release|x86.Build.0 = Release|Any CPU
 		{57B6F9D9-C410-4EB7-8003-60F894F358A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{57B6F9D9-C410-4EB7-8003-60F894F358A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{57B6F9D9-C410-4EB7-8003-60F894F358A4}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
