--- RadioOverlay/App.xaml
@@ -1,15 +1,0 @@
-﻿<Application x:Class="Ciribob.DCS.SimpleRadio.Standalone.Overlay.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:Ciribob.DCS.SimpleRadio.Standalone.Overlay"
-             StartupUri="MainWindow.xaml">
-    <Application.Resources>
-
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="Themes\Styles.xaml" />
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-
-    </Application.Resources>
-</Application>
