--- RadioOverlay/App.xaml.cs
@@ -1,11 +1,0 @@
-﻿using System.Windows;
-
-namespace Ciribob.DCS.SimpleRadio.Standalone.Overlay
-{
-    /// <summary>
-    ///     Interaction logic for App.xaml
-    /// </summary>
-    public partial class App : Application
-    {
-    }
-}
