--- RadioOverlay/Themes/LeftMarginMultiplierConverter.cs
@@ -1,31 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-
-namespace DarkBlendTheme
-{
-    public class LeftMarginMultiplierConverter : IValueConverter
-    {
-        public double Length { get; set; }
-
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            var item = value as TreeViewItem;
-            if (item == null)
-                return new Thickness(0);
-
-            return new Thickness(Length * item.GetDepth(), 0, 0, 0);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new System.NotImplementedException();
-        }
-    }
-}
