--- RadioOverlay/packages.config
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-
-<packages>
-  <package id="Costura.Fody" version="1.3.3.0" targetFramework="net452" developmentDependency="true" />
-  <package id="Fody" version="1.29.4" targetFramework="net452" developmentDependency="true" />
-  <package id="NBug" version="1.2.2" targetFramework="net452" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net452" />
-  <package id="NLog" version="4.3.5" targetFramework="net452" />
-  <package id="WPFThemes.DarkBlend" version="1.0.8" targetFramework="net452" />
-</packages>
